requirements.txt
Flask==3.1.0
ImageHash==4.3.2
imbalanced_learn==0.13.0
imblearn==0.0
instaloader==4.14.1
joblib==1.4.2
numpy==2.2.3
pandas==2.2.3
Pillow==11.1.0
Requests==2.32.3
scikit_learn==1.6.1
scipy==1.15.2
setuptools==65.5.0
xgboost==2.1.4
